# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Comprehensive REST API for managing game server infrastructure with panel integration, async job queue management, and real-time sync operations",
        "title": "NodeByte API",
        "termsOfService": "https://nodebyte.co.uk/legal/terms",
        "contact": {
            "name": "Contact Support",
            "url": "https://discord.gg/wN58bTzzpW"
        },
        "license": {
            "name": "AGPL 3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "0.2.0"
    },
    "host": "core.nodebyte.host",
    "basePath": "/",
    "paths": {
...
        "/api/v1/hytale/oauth/device-code": {
            "post": {
                "description": "Initiates OAuth 2.0 Device Code Flow for Hytale server authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale OAuth"
                ],
                "summary": "Request Device Code",
                "parameters": [
                    {
                        "description": "Device code request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.DeviceCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeviceCodeResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hytale/oauth/game-session/delete": {
            "post": {
                "description": "Terminates an active game session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale OAuth"
                ],
                "summary": "Terminate Game Session",
                "parameters": [
                    {
                        "description": "Terminate game session request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.TerminateGameSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.TerminateGameSessionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hytale/oauth/game-session/new": {
            "post": {
                "description": "Creates a new game session for the selected profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale OAuth"
                ],
                "summary": "Create Game Session",
                "parameters": [
                    {
                        "description": "Create game session request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateGameSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateGameSessionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Session limit reached",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Token or profile not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hytale/oauth/game-session/refresh": {
            "post": {
                "description": "Refreshes an existing game session to extend its lifetime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale OAuth"
                ],
                "summary": "Refresh Game Session",
                "parameters": [
                    {
                        "description": "Refresh game session request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RefreshGameSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.RefreshGameSessionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/hytale/oauth/profiles": {
            "post": {
                "description": "Fetches available game profiles for the authenticated account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hytale OAuth"
                ],
                "summary": "Get Game Profiles",
                "parameters": [
                    {
                        "description": "Get profiles request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GetProfilesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetProfilesResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Token not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.DeviceCodeRequest": {
            "type": "object",
            "properties": {
                "account_id": {
                    "description": "Account/Owner UUID from Hytale",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "types.DeviceCodeResponseDTO": {
            "type": "object",
            "properties": {
                "device_code": {
                    "type": "string",
                    "example": "AH-wO0aD5nvS5xhd7rQw1qv6XUzC9Kk9IElVqxsqQ1KGIykN3iqjcqB5hFtMQxPuBs4uEA"
                },
                "error": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer",
                    "example": 900
                },
                "interval": {
                    "type": "integer",
                    "example": 5
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "user_code": {
                    "type": "string",
                    "example": "GFHD-MJHT"
                },
                "verification_uri": {
                    "type": "string",
                    "example": "https://accounts.hytale.com/oauth2/device?user_code=GFHD-MJHT"
                },
                "verification_uri_complete": {
                    "type": "string",
                    "example": "https://accounts.hytale.com/oauth2/device?user_code=GFHD-MJHT"
                }
            }
        },
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Human-readable error message",
                    "type": "string",
                    "example": "account_id is required"
                },
                "success": {
                    "description": "Whether the request was successful",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Backend-to-backend API authentication",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}